extract editor extensions to react hooks
